pyproject.toml
[tool.poetry]
name = "CPA-Network"
version = "0.1.0"
description = ""
authors = ["AsL <bloogrox@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4"
celery = "5.2.7"
"dj-database-url" = "0.5.0"
Django = ">=4.2,<4.3"
"django-cors-headers" = "3.11.0"
"django-countries-plus" = "2.2.0"
"django-filter" = "21.1"
"django-rest-swagger" = "2.2.0"
"django-tinymce" = "3.5.0"
djangorestframework = "3.14.0"
"djangorestframework-simplejwt" = "*"
"drf-spectacular" = "0.27.0"
gunicorn = "20.1.0"
honcho = "1.1.0"
"psycopg2-binary" = "2.9.5"
redis = "4.3.5"
requests = ">=2.32"
whitenoise = "6.2.0"
iso8601 = "^2.1.0"
webargs = "^8.7.0"


[tool.poetry.group.dev.dependencies]
"django-stubs" = "1.13.0"
"djangorestframework-stubs" = "1.8.0"
mypy = "0.991"
pytest = "*"
ruff = "*"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
exclude = ["venv", ".git", "__pycache__", "migrations"]
line-length = 79
select = [
    "E",   # pycodestyle rules
    "F",   # Pyflakes rules
    "C90", # McCabe complexity
]

[tool.ruff.mccabe]
max-complexity = 10
